+++ packages/CodingStandard/tests/Sniffs/ControlStructure/SprintfOverContact/SprintfOverContactSniffTest.php
@@ -1,0 +1,43 @@
+<?php declare(strict_types=1);
+
+namespace Symplify\CodingStandard\Tests\Sniffs\ControlStructure\SprintfOverContact;
+
+use Iterator;
+use Symplify\EasyCodingStandardTester\Testing\AbstractCheckerTestCase;
+
+/**
+ * @see \Symplify\CodingStandard\Sniffs\ControlStructure\SprintfOverContactSniff
+ */
+final class SprintfOverContactSniffTest extends AbstractCheckerTestCase
+{
+    /**
+     * @dataProvider provideWrongCases()
+     */
+    public function testWrong(string $file): void
+    {
+        $this->doTestWrongFile($file);
+    }
+
+    public function provideWrongCases(): Iterator
+    {
+        yield [__DIR__ . '/wrong/wrong.php.inc'];
+    }
+
+    /**
+     * @dataProvider provideCorrectCases()
+     */
+    public function testCorrect(string $file): void
+    {
+        $this->doTestCorrectFile($file);
+    }
+
+    public function provideCorrectCases(): Iterator
+    {
+        yield [__DIR__ . '/correct/correct.php.inc'];
+    }
+
+    protected function provideConfig(): string
+    {
+        return __DIR__ . '/config.yml';
+    }
+}